+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/application/ScheduleRestController.java
@@ -32,4 +32,9 @@ public class ScheduleRestController {
     public List<Schedule> getScheduleByCourse(@PathVariable Long courseId){
         return scheduleController.getScheduleByCourse(courseId);
     }
+
+    @GetMapping ("/class/{classId}")
+    public List<Schedule> getScheduleByClass(@PathVariable Long classId){
+        return scheduleController.getScheduleByClass(classId);
+    }
 }