+++ NIMIO2024_TFG_PolMorenoCastelar/changelog.md
@@ -101,9 +101,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - [NPP-31] 》SC23F - Get schedule by classId
 
-    ## Schecule Extra Functionalities
+    ## Schedule Extra Functionalities
     
     - [NPP-68] 》EX54F - Get all schedules for a course by courseID
+    - [NPP-67] 》EX53F - Get all schedules for a class by classID